requirements.txt
textual>=0.80
pytest-asyncio>=0.23
